# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI Course 后端 API 服务",
        "title": "AI Course API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/assignment": {
            "post": {
                "description": "教师创建新作业",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "作业管理"
                ],
                "summary": "创建作业",
                "parameters": [
                    {
                        "description": "作业信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateAssignmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/assignment/list": {
            "get": {
                "description": "获取当前教师的所有作业列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "作业管理"
                ],
                "summary": "获取教师作业列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/assignment/{id}": {
            "get": {
                "description": "获取作业详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "作业管理"
                ],
                "summary": "获取作业详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "作业ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "作业不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "教师更新作业信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "作业管理"
                ],
                "summary": "更新作业",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "作业ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "作业信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateAssignmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "作业不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "教师删除作业",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "作业管理"
                ],
                "summary": "删除作业",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "作业ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "作业不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/assignment/{id}/publish": {
            "post": {
                "description": "教师发布作业",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "作业管理"
                ],
                "summary": "发布作业",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "作业ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "发布成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "作业不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/assignment/{id}/statistics": {
            "get": {
                "description": "教师获取作业统计信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "作业管理"
                ],
                "summary": "获取作业统计信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "作业ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "作业不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/assignment/{id}/unpublish": {
            "post": {
                "description": "教师取消发布作业",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "作业管理"
                ],
                "summary": "取消发布作业",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "作业ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "取消发布成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "作业不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CreateAssignmentRequest": {
            "type": "object",
            "required": [
                "class_id",
                "deadline",
                "title"
            ],
            "properties": {
                "class_id": {
                    "type": "integer"
                },
                "deadline": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CreateQuestionRequest"
                    }
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "draft",
                        "published"
                    ]
                },
                "title": {
                    "type": "string",
                    "maxLength": 200
                },
                "total_score": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "model.CreateQuestionRequest": {
            "type": "object",
            "required": [
                "content",
                "order",
                "score",
                "type"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "correct_answer": {
                    "type": "string"
                },
                "explanation": {
                    "type": "string"
                },
                "is_multiple": {
                    "description": "选择题是否多选",
                    "type": "boolean"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.QuestionOption"
                    }
                },
                "order": {
                    "type": "integer",
                    "minimum": 1
                },
                "reference": {
                    "type": "string"
                },
                "score": {
                    "type": "integer",
                    "minimum": 1
                },
                "type": {
                    "enum": [
                        "choice",
                        "fill_blank",
                        "true_false",
                        "essay"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.QuestionType"
                        }
                    ]
                }
            }
        },
...
        "model.QuestionOption": {
            "type": "object",
            "properties": {
                "key": {
                    "description": "选项标识 A, B, C, D",
                    "type": "string"
                },
                "value": {
                    "description": "选项内容",
                    "type": "string"
                }
            }
        },
        "model.QuestionType": {
            "type": "string",
            "enum": [
                "choice",
                "fill_blank",
                "true_false",
                "essay"
            ],
            "x-enum-comments": {
                "QuestionTypeChoice": "选择题",
                "QuestionTypeEssay": "简答题",
                "QuestionTypeFillBlank": "填空题",
                "QuestionTypeTrueFalse": "判断题"
            },
            "x-enum-varnames": [
                "QuestionTypeChoice",
                "QuestionTypeFillBlank",
                "QuestionTypeTrueFalse",
                "QuestionTypeEssay"
            ]
        },
...
        "model.UpdateAssignmentRequest": {
            "type": "object",
            "properties": {
                "deadline": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "draft",
                        "published",
                        "closed"
                    ]
                },
                "title": {
                    "type": "string",
                    "maxLength": 200
                },
                "total_score": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应信息",
                    "type": "string"
                }
            }
        },
...
    }
}